finalize docker
